+++ Mainboard_ESP32/components/com/lora.h
@@ -1,3 +1,4 @@
+// Copyright 2023 PWr in Space, Krzysztof Gliwiński
 #pragma once
 
 #include <stdint.h>
@@ -73,14 +74,14 @@
  * \param reg Register index.
  * \param val Value to write.
  */
-void lora_write_reg(int reg, int val);
+void lora_write_reg(int16_t reg, int16_t val);
 
 /*!
  * \brief Read the current value of a register.
  * \param reg Register index.
  * \return Value of the register.
  */
-int lora_read_reg(int reg);
+int16_t lora_read_reg(int16_t reg);
 
 /*!
  * \brief Perform physical reset on the Lora chip
@@ -98,7 +99,7 @@ void lora_explicit_header_mode(void);
  * All packets will have a predefined size.
  * \param size Size of the packets.
  */
-void lora_implicit_header_mode(int size);
+void lora_implicit_header_mode(int16_t size);
 
 /*!
  * \brief Sets the radio transceiver in idle mode.
@@ -122,43 +123,43 @@ void lora_receive(void);
  * \brief Configure power level for transmission
  * \param level 2-17, from least to most power
  */
-void lora_set_tx_power(int level);
+void lora_set_tx_power(int16_t level);
 
 /*!
  * \brief Set carrier frequency.
  * \param frequency Frequency in Hz
  */
-void lora_set_frequency(long frequency);
+void lora_set_frequency(int32_t frequency);
 
 /*!
  * \brief Set spreading factor.
  * \param sf 6-12, Spreading factor to use.
  */
-void lora_set_spreading_factor(int sf);
+void lora_set_spreading_factor(int16_t sf);
 
 /*!
  * \brief Set bandwidth (bit rate)
  * \param sbw Bandwidth in Hz (up to 500000)
  */
-void lora_set_bandwidth(long sbw);
+void lora_set_bandwidth(int32_t sbw);
 
 /*!
  * \brief Set coding rate
  * \param denominator 5-8, Denominator for the coding rate 4/x
  */
-void lora_set_coding_rate(int denominator);
+void lora_set_coding_rate(int16_t denominator);
 
 /*!
  * \brief Set the size of preamble.
  * \param length Preamble length in symbols.
  */
-void lora_set_preamble_length(long length);
+void lora_set_preamble_length(int32_t length);
 
 /*!
  * \brief Change radio sync word.
  * \param sw New sync word to use.
  */
-void lora_set_sync_word(int sw);
+void lora_set_sync_word(int16_t sw);
 
 /*!
  * \brief Enable appending/verifying packet CRC.
@@ -173,14 +174,14 @@ void lora_disable_crc(void);
 /*!
  * \brief Perform hardware initialization.
  */
-int lora_init(void);
+int16_t lora_init(void);
 
 /*!
  * \brief Send a packet.
  * \param buf Data to be sent
  * \param size Size of data.
  */
-void lora_send_packet(uint8_t *buf, int size);
+void lora_send_packet(uint8_t *buf, int16_t size);
 
 /*!
  * \brief Read a received packet.
@@ -188,17 +189,17 @@ void lora_send_packet(uint8_t *buf, int size);
  * \param size Available size in buffer (bytes).
  * \return Number of bytes received (zero if no packet available).
  */
-int lora_receive_packet(uint8_t *buf, int size);
+int16_t lora_receive_packet(uint8_t *buf, int16_t size);
 
 /*!
  * \returns non-zero if there is data to read (packet received).
  */
-int lora_received(void);
+int16_t lora_received(void);
 
 /*!
  * \returns last packet's RSSI.
  */
-int lora_packet_rssi(void);
+int16_t lora_packet_rssi(void);
 
 /*!
  * \returns last packet's SNR (signal to noise ratio).
@@ -211,6 +212,6 @@ float lora_packet_snr(void);
 void lora_close(void);
 
 /// Not supported
-int lora_initialized(void);  
+int16_t lora_initialized(void);
 
 void lora_dump_registers(void);
